+++ server/src-lib/Hasura/App.hs
@@ -385,7 +385,7 @@ waitForShutdown = C.takeMVar . unShutdownLatch
 -- | Initiate a graceful shutdown of the server associated with the provided
 -- latch.
 shutdownGracefully :: ShutdownLatch -> IO ()
-shutdownGracefully = flip C.putMVar () . unShutdownLatch
+shutdownGracefully = void . flip C.tryPutMVar () . unShutdownLatch
 
 -- | If an exception is encountered , flush the log buffer and
 -- rethrow If we do not flush the log buffer on exception, then log lines